fix: add signup action callback
